use preinstalled-pool feature to provider bootloaders for offline installs
this is a feature of auto/build of livecd-rootfs. specifically our fork
has a special adjustment to this feature from ubuntu.

- in a config hook we'll grab overrides (for prosperity), this also
  creates a bunch of dirs auto/build will expect to be there
- in a build hook we'll then download a bunch of packages into
  a special preinstalled-pool directory, auto/build will find this and then
  generate a proper repo out of it and add it to the ISO's sources.list
- our livecd-rootfs-neon fork has an adjustment to this behavior, where
  ubuntu adds the pool to the sources.list we add them to sources.list.d/
  rationale being that ubiquity rewrites sources.list, so this would get
  lost during installation, while using sources.list.d/ allows us to
  preserve this into the installation making sure grub-installer works
  as expected

NB: the downloaded packages have no dep tracking, we maybe should change to
  `install -d` and move the debs around or add some post-build testing to
  make sure the packages actually work (e.g. bcmwl-kernel-source doesn't
  and never did in xenial, begging the question why it is getting packed
  at all)

the list of packages comes from the dumped repo we have as a copy in
live-build (gets added as cdrom repo)

this is slightly less nice than the cdrom repo which wouldn't end up on the
final disk, but in the grand scheme of things it matters little. I also
couldn't find what exactly generates the cdrom repo, so meh.