add userns awareness to workspace cleanup
if docker is in userns mode we need to chown to root:root rather than
jenkins:jenkins as userns remaps containerRoot == hostUser and all uids
in the container are offset from that. so we have hostUser 10000 and
subuid 10000 and then containerUid 2 is hostUid 10002

technically we should probably chown by uid rather than name though. that
is still off set however, so it serves no real purpose as we'd still have
to handle the unoffseting