disable seccomp confinement
qt 5.10 uses statx, statx is not whitelisted and cannot be whitelisted
presently. this is a workaround for deficiencies in docker and really
should not be necessary :(

generally all stuff we run should be sufficiently trusted. also, for
most slaves the containers have user-level permissions on the host already
so there's a natural limit to the syscall permissions already.