new feature for watch fetcher: grab from archive if series is given
there is a problem with how our sourcers work and how the archive
works. each series has its own sourcer, but on an archive level
they all have the same pool. now if a sourcer repacks the tarballs
generated may be different between e.g. xenial and bionic.
this would then result in a rejected source upload as the tarballs
would have a different checksum.

to make this a less likely scenario the uscan fetcher now can
attempt to grab a tarball from the archive as a first choice.
only if it can't find a suitable one it will actually invoke uscan
and thus potentially repack.

this still leaves a reentrancy problem where two series may build
the same source at the same time and both repack and then both
fail on pub. there isn't much we can do about this I fear, but
at least a simple retry should fix this the second time around as
one of the two would  have successfully published and this tarball
can now be used by the other job on a retry.

NB: this also requires the apt sources to be configured correctly