automatically let sourcers resolve dependencies
this still preserves the original behavior of assuming the deps are in
the image. however, if the source build fails we'll attempt to "fix" it
automatically by installing the deps on-demand. if it fails again we'll
error out at this point

this helps arbitrary !kde packages build without a human blowing up the
core image with stupid build depends only used in a single package