fix source adding to work for !current series
apt has a buggy where it won't let you `apt-get source -t` if the t
can't be found in the package cache. it can only be found in there if
it actually had deb sources enabled though. as a stop gap we'll simply
add the !current deb entry as well.
since we already set the current series as default-release (thus forcing
its preference over other available ones) this should be fine to use as
the right series is preferred even for the deb packages

(does make the cache faster though)