drop throttle plugin
it has abysmal performance impact on the jenkins Queue's maintain() method
which in turn means everything gets super slow as the Queue has fairly
sweeping locks in place. everything that needs to poke the queue (
which is a lot of things) will have resource contention as a result of this
as locks are held for way too long when iterating the task list.
the more tasks are in the queue the worse this gets as basically each
maintain() will call potential plugins extensions into the queuing system
which in turn will almost all iterate the entire list of tasks and do
non trivial operations on them (e.g. a whole bunch of string compares).

the throttle plugin seems particular bad in that regard with its
canTakeImpl() looking at the very least O(n) and doing a whole lot of
non trivial compares on all tasks and all pipelines

dropping the throttle plugin seems to reduce the overall contention quite
a bit, however maintain() still gets called too often when doing mass
operations causing occasional oversubscription, but certainly not to the
end that the throttle plugins seems to do.

(currently heaviest plugin, with throttle removed, seems to be dependency
 queue, which we actually need. might be as good as it gets I fear)