fix tarball locating to more aggressively discard shitty tarballs
someone somehow managed to get the qbs source cache littered with random
stuff of various versions at some point. this went good for a while
but now screws up tarball finding.

the new logic restricts the tarball to not the newest but the exact version
we are looking for

update a text fixture for which this was not the case (matched newest, not
what was in changelog)