implement a safety mechansim to prevent repacks on !xenial for nci
the uscan sourcer will try to get the tarball from apt when provided with
a list of series to fetch from. however, when both xenial and bionic
build at the same time both would fail to find a tar and both would repack
one and both would eventually attempt to upload (the divergent) tarballs
resulting in fatal errors inside aptly.

to guard against this situation we'll flat out refuse to uscan repacking
tarballs on the !current nci series (i.e. !xenial) and raise an error
instead.

once the xenial build is done the source ought to be findable and the guard
would pass as now apt source should find a tarball