try to roll back part of the iso
on reboot we now get no plymouth but instead a whole bunch of squashfs
errors.

however, I can confirm that the squashfs errors happen AFTER the hit
enter to reboot message.

so it may be that the errors aren't new and we the problem is that plymouth
simply isn't starting.
OR the problem is that the errors are new and that is why plymouth
doesn't start.

roll back parts of the ISO to versions we were on before to figure out
what broke it. (manually respinning parts of the iso on my system
resulted in plymouth always starting... this is the stupidest bug...)