Revert "remove preinstall pool enablement"
This reverts commit d6eafd9ab1cf6d226be4aeb042b1f3ea3d977a81.

attempt to bring back the preinstall pool to debug what went wrong there.
bionic uses gpg2 by default, with gpg1 being available as opt-in.

however I think this the pertinent code in livecd-rootfs runs on the host
(so xenial, so gpg1) so it may not be a gpg2 compat problem.