use stable frameworks for candidate snaps
so... the idea is to validate the snap against the stable frameworks.
there will be a second test of set which validates a candidate framework
against a stable snap, and there may well be both candidate.

the purpose of this specific job is to assert that the snap is in candidate
is working well enough to get promoted to stable. to establish that we
should not take candidate frameworks into account. for one they may
have regression of their own, for another if the ABI in stable isn't good
enough for this snap we want the QA to fail