assert that all tests are using our testcase instead of test-units'
originally some of the older tests were left on test-unit's TestCase
as they had no reason to use our testcase variant (as in: they were
meta tests unrelated to any of our code or fixture tech)

however, Jonathan either doesn't know or understand that we want our
TestCase for everything doing non-trivial tests. so, in the interest
of making that requirement explicit and obvious we now assert that in fact
all tests are using the right class.

as a result of this I've moved all of the meta tests to testcase as well,
given it's supreme metaprogrammingness there is no longer a downside
to using our TestCase anyway