prevent canceled binary jobs from causing bogus publishing
when a bin is canceled (i.e. unqueued), the multijob gets a
cancelexception which isn't correctly handled (as in: ignored). if we then
proceed to the publishing step we'll publish only a partial build
(e.g. only armhf debs), the repo cleaner will then drop the old build
(with all architectures) from the repo and we are left with a crippled repo
that only has armhf for the source.

to prevent this the bin jobs now re-archive the run_stamp. this means
all bins regardless of them building something or not will have at least
one artifact the publisher can unarchive.
this then allows us to switch the publisher to not consider the bin
unarchiving optional.

the end result is that if a bin job didn't run at all the publisher will
fail to unarchive that bin job and fail, which in turn causes the multijob
as a whole to fail