put daily promote onto cloud nodes
this avoids out of disk errors with persistent nodes as they are notorious
for leaking. also, concurrent large builds from dci can easily push
the nodes over the limit.

this is using regular build nodes for now, technically it'd probalby be
more appropriate to spin up a special type only for this, but in the grand
scheme of things it likely matters not at this particular job has no
real chance of leaking and leakage from actual build jobs should also
be small enough to not then cause disk space issues on the cloud nodes.

(should this be a problem in the future obviously this simply needs moving
onto a special node type only used for these jobs; much like we do for
iso building)