refactor the openqa install env to always have "core" variables
this is to facilitate moving away from the many jenkinsfiles. by fully
controlling the parameterization of the jobs via the environment property
we can use the same jenkinsfile and create permutations through only the
env. this relies on a pending change in the openqa repo

as a result the script_path overrides in the secureboot and oem jobs
have been replaced by env overrides

also, to prevent accidently overriding env and losing the global
INSTALLATION var, the init of the base job now asserts that the env is
from the base job. derivates are meant to override exttra_env, which gets
appended to env.