add a new series_restrictions array to project instances
this member is representing a restriction to 0:N series instead of all.
the purpose of this is to allow restricting any amount of projects to
e.g. build on xenial but not bionic or vice versa.

like series_branches this entirely depends on the project_updater impls
to actually implement support for the restriction

this commit on its own is not functionally useful as there is no way to
set the restrictions manually. restrictions are (so is the plan) only set
by overrides. meaning a project_updater needs to load CI dependent
overrides before factorizing the projects. while the overrides code for
this has been there since forever we previously had no use for it. tldr:
project_updaters also need updates to actually load overrides to set
the restrictions IN ADDITION TO having to implement the restriction itself

this commit also adds the ability for rules to override basic types (!Hash)
previously we would only ever apply overrides to objects as their instances
while now we can also set basic types directly.

because of how overrides cascade multiple files may be loaded to construct
a complete series_restrictions array. on the flip side you can't actually
remove a restriction inserted by another override rule. should this ever
be needed in the future we'll need to revisit this. possibly with a
custom data type instead of array