implement projects' series_restrictions support in nci job updater
series_restrictions allow us to restrict projects to specific series of
neon. e.g. a job may only apply to xenial (NCI.current_series) if it is
a backport, in that case we don't want jobs for it for bionic
(NCI.future_series)

for prosperity we'll log when a project gets kipped because of a
restriction