allow lunchpad to be overridden
launchpad has no concept for branches inside repos, instead repos are
branches. to still support overrides we'll simply treat nil branches as
empty branches. we need a stringy here so the pattern class doesn't fall
over dead, and the empty string seems fairly suitable to present no branch.

(this could also be done as a separate format for launchpad, but in the
interest of format consistency it's probably best to not do that)

this enables overriding bzr things via branch rules '*' or ''