refactor snapshot jobs to differentiate type by class
the keywording of this job type is intense. wrap it out of the job
updater and into type classes for release and lts. this way the job
updater is lots easier to read.

NB: the kwords basically just moved into the wrapper classes. they could
probably be undone largely by smarter logic but I didn't want to be to
invasive with this change