add a cheeky release upgrader publisher hack
release upgrader builds a tarball which is meant to get published somewhere
separately from the debs. the fetcher then downloads the tarball and runs
the upgrader in the tarball. ubuntu does that in a regular metadata dir
in the repo. so, let's do the same.

the helper does our usual pushing crap by sshing into the suitable host,
uploading assets, signing them on the remote and the moving them to the
aptly skel dir

because of the skelling this is unfortunately somewhat hard to include in
snapshot tech.

this is only active for release builds. in part because of skelling, but
also because all builds are the same anyway, so release is as good as
any.

code could use some testing but I feel like I need to write a mock ssh
session class for use in multiple tests at this point, all our pushes
aren't really tested and all of them because of the mock overhead