do not force remove running containers
this behavior is fairly ancient and pre-dates the pangea namespace
check.

this is a bit of a balancing act. we want to remove as many dangling
containers as possible but not remove containers a human may be using or
that may be used to serve a persistent service.
to achieve that the namespace check was added which would only remove
containers iff they belong to our namespace. unfortunately that was
undermined by running containers getting cleaned with force:true.
the new behavior is to only force exited old containers. running containers
are only removed if they are based on pangea images

this is now also being a bit debugged so we can reproduce why containers
got removed. the conditional for aborting the remove a container was also
made easier to read