raise a runtimeerror with context on jenkins template render errors
previously we'd get useless errors such as "no method foo on nil" or
whatever else the error in ERB was. the problem is that the callchains are
entirely generic from updater.rb into template.rb, so we'd have no
clue what actual job caused the error or what its state was.
to ease debugging of template errors we'll now catch all standard errors
and re-raise them as runtime errors with context of the object and its
internal state attached.