iterate architectures + stash without subdirs
this adds programtically generated parallel stages for the various
architectures we may want instead of hardcoding them.

this additionally chdirs before stashing. this allows us to get rid of the
subdirs in the stashes and means we can drop all the subdir handling code
(this was also responsible for no uploads actually happening during
 publishing as I didn't even think to handle the subdirs there)