correctly derive sourcing type from upsteam_scm [untested]
the sourcing type was incorrectly derived from the packaging_scm, we do
however need it derived from the upstream_scm (as that may be git/uscan/
tarball)

the new code lonely operates its way to type. this should correctly end up
nil in the event of upstream_scm not being set at all (native pkg) which
in turn will render an empty string in the pipeline. this should be
exactly how SourcerJobs do this