always enable dh_missing list missing via a bin overlay
dh_missing list missing would ordinarily only run with compat 12 (
which isn't even a thing in bionic yet).
so in the interest of getting good behavior now we'll always run
list missing on dh_missing.

should this cause troubles we'll have to roll back and figure out some
workarounds. the use greatly outweighs the annoyance of having to do
possible workarounds for bugs in dh