fix excessive asgen rebuilds
when causing a configure call with meson it will prompt a rebuild of
everything. even when nothing has changed!!!!
to prevent this we'll only configure once, ever. the only way to
reconfigure is to wipe the build dir or add new code.

this also moves away from chdiring in the ruby side, since it's nicer to
instruct the tools on doing the chdir (thread-safe on our side)