excessively revise the way asgen builds to prevent rebuilds
meson continues to be daft and rebuilds when it shouldn't. this new
approach writes the git rev we've built last into file. on subsequent
runs we'll read the file to determine whether to even install builddeps
and run any build commands.

this should prevent any rebuilds unless the git rev changes