run asgen on all repo types
this moves away from the legacy rysnc based approach to a modern
on-demand download as supported by asgen. this currently relies on
a personal branch I have as the original code in bionic is broken for
our repos and a fix from proper master is difficult to build because the
build system of master has moved so far along that it can't build on
bionic anymore.

- drops old asgen bins
- new asgen bin in main nci dir (instead of ci-tooling)
 - for improvements on that asgen see previous commits
- jobs are now generated for all repos and name was adjusted accordingly