skip all targets that have Qt5::Test as link target
this is a poor man's check for test targets, unfortunately we have no
reasonable way to determine if a target is for a test or not. test targets
have no special targets unless they get defined with one of the test
specific properties. also cmake has no facility to list all targets until
cmake 3.12 (which is too new for us)