snapcraft extender now asserts DIST == base
this should also happen elsewhere, but since extender already has multibase
support its cheapest to add this here.

this simply asserts that the DIST (i.e. the var which determined what
base docker image to load) is the same as the base version of the snap
to be built. if they don't add up then one tries to build a snap against
the wrong system ABI