add new tech to be able to deal with build-snaps
build-snaps are basically dev packges as snaps, problem is we'd need
a snapd to actually use them properly. currently we have no setup to
properly deal with snapd inside our dockers, so we'll pull a trick seen
in snapcraft's own bootstrap code.
instead of relying on snapd we'll manually get the snap file and unpack
it into place. i.e. we are basically mocking snapd's behavior here.
alas, we also need to mock some of snapcrafts (future) behavior as we'll
need to also take over some magic that it would do when using build-snaps.
I say future because this feature has not been landed, but is under review