fix collapsed to properly only support cmake
previously we'd always attempt to add confligflags, when in fact we didnt
need them on account of not having root paths, and we'd not have root paths
if no build-snap is set and build-snap may only be set for cmake
(right now). this way of dealing with things should be relatively future
proof. should we find that supporting more types is necessary we'd also
have to rejigger the confligflags logic anyway