merge content and dev stage list for exclusion if a -dev package is staged
this is mostly to support shortcut stage-packages where we only list
the -dev and that *should* drag in the libs we want.
this may actually need some extra tech in the long term where the snap
plugin handles this properly and pulls in everything but then doesn't
consider the -dev as a stage deb but rather its dependencies

anyway. good for now I think