add build profile support in deb822 relationships
new attribute profiles on Relationship which is an array of ProfileGroups
(which is derived from Array), each Group contains 1:N Profile instances
a Profile is an identifier string plus a possible negation

ProfileGroups and Profiles have a matches? method to evaluate an input
for applicability to a given profile(group)

since this is horribly complex to use there is a single frontend helper
method on Relationship called #applicable_to_profile? which wraps the
task of evaluating if the Relationship is applicable to any input profile.
specifically this can be used like
```
rel.applicable_to_profile?(ENV['DEB_BUILD_PROFILES'])

```
to check if a relationship is to be used under the current environment