extend the manifest extender with a way of dealing with build deps
snapcraft build-package are not broken up by manifest.txt, so we need
a different way of dealing with them. we'll simply install fake packages
for all the packages we have in the content-snap/build-snap. thus allowing
us to get away with not installing stuff we already have (or such is
the intent anyway)