drop legacy bind behavior
originally pwd was bound under pwd inside the container. this was changed
long ago to instead be controlled via PWD_BIND and kind-of standardized
as /workspace

since this has been a todo since forever let's just jump into the deep end
and drop the legacy code. if stuff breaks we can easily fix it up I expect

the new behavior always set the bind target to /workspace for convenience
but still follows PWD_BIND as a means to override that