implement persistent cache storage
we'll now store persistent data on the master by stashing it around between
slave and master. I feel like we need a faster but as convenient solution
as rsync at some point. stash isn't scaling too well.

so. this is a bit complicated. the stash does explicitly not include
the export/ dir, the reason it doesn't is because it makes everything
faster. we don't need to stash the export (which is already stored via
metadata.neon.kde.org) so that's fast, and asgen doesn't generate new
export data when the data didn't change, so we have an easy way to
determine if there was a change or not. when there was no change we can
skip all of publish and be even faster. hooray!