unpack by-hash for dep11 [experimental; hackish]
needs some refactor + testing still

aptly currently doesn't generate by-hash for skel files (i.e. the way
we glue dep11 into the repo metadata. as a mid term stop-gap we'll manage
by-hash manually via asgen_push

this is a bit fiddly and hackish but should be working reasonable enough.
before pushing we'll checksum all files, then sync a (possibly)
pre-existing by-hash dir, and adjust its content accordingly. this is
using a similar approach to aptly. the current variant is Foo which
symlinks to 12345 and that is a copy of Foo right now, meanwile Foo.old
is a symlink to 9876 which is a copy of the previously published Foo.
this effectively means we always have two variants by-hash. this could
actually benefit from some more, but we can't control this on aptly
currently and only having more for dep11 doesn't do anything much.

push needs refactoring into class so we can unit test this. the by-hash
tech is a fair amount of code which isn't particularly straight-forward