switch to wrapers to reduce code duplication and move to any amd64
this was unjustly sent to the cloud. particularly with the new code
spin up of the cloud can take longer than the actual generate step, so
instead go for any available amd64 node. even the persistent ones
will be fast enough to deal with this

also, the actual asgen portion is now wrapped in cleanup so as to not
leak workspaces on the peristent nodes