don't report setcap violations if the sub process (i.e. build) failed
people are misreading the output when the first error they see (from the
end) is setcap.

instead introduce a new setcaperror, raise that error instead of a generic
runtimeerror. tests assert that actual violations raise setcaperror while
build failures a runtimeerror

the run ensure now only stops the setcap helper but doesn't actually
assert any results. this ultimately can be a double-stop if no error
was raised. that seems to be save and is covered by the success test so
it won't be causing problems in the future either