don't trip over leading whitespaces when searching dh_missing
the dh_missing code is special in that regexes for the relevant segment
to parse, this is a bit problematic as apparently sometimes the log
can have weird leading whitespaces in the pertinent lines.
to bypass this problem we'll allow leading space characters in our
regex.
additionally the lines get stripped before parsing to ensure they too
won't fall over unepxected whitespaces