add opt-in hack to install snapcraft snap
this uses the same hacky approach as we use for the build snap by
downloading the snap and manually unsquashing it into place. since
snapcraft is a classic snap it's rpath'd and thus able to run on its own
even without additional help from snapd.

we do however still need a wrapper to setup some essential env stuff.
luckily upstream also has that for their testing in docker.

https://phabricator.kde.org/T10834