abstract the qt staging type name
this is both to not have to go around changing code should the type
name change again in the future. but also to make its uses slightly more
readable in that it'll be obvious the qt stage is excluded. a fact which
is not immediately obvious with the type name itself as one would know
what that type refers to