make projectupdater only install plugins that are actually used
previously we'd compile a list of plugins from the available templates
(which is technically faster) because of how Rohan stacked the templates
in the first refactor to support multiple CIs this did however also mean
that each CI would install plugins in templates that were in the shared
dir. while that dir has grown more and more empty over time, the install
behavior is still incorrect. this is particularly a problem with nci
where ircbot isn't meant to be used anymore yet the shared templates
would continue to pull it in

fix the problem by detecting used plugins in each job rather then the
base templates. this is slower because it needs to eval some 1k jobs, but
more accurate because in fact only used stuff will get installed