fix loop detection
previously we'd only find loops that go back to the original job, instead
we now collect a stack and check that we do not loop inside the stack.

this covers the previous case where we'd go back to the root job, but also
cases where any of the intermediate steps loops back in on itself

e.g.
A -> B -> C -> B