force-lower versions on experimental builds
since we want to copy packages from experimental to unstable we need to
make sure that the version in experimental cannot have a greater build
number than unstable as the copied build would then forever exceed the
unstable version.

this commit adds a ~exp qualifier just before the build qualifier to force
it to be always lower than a build from unstable IFF the preceeding
qualifiers are the same

e.g.
5.13.1-0xneon+18.04+bionic~exp+build5 (experimental)
<<
5.13.1-0xneon+18.04+bionic+build3 (rebuild of experimental in unstable)

BUT

5.13.1-0xneon+18.04+bionic~exp+build5 (experimental)
>>
5.13.0-0xneon+18.04+bionic+build2 (old qt in unstable)