new s/archive/expunge job for nci
archiving is no longer a thing since we have no "archive" disk so to speak
instead move to a full expunging system for that purpose the job is now
renamed and the scripty as well and has entirely simpler functionality
(which is why I don't bother writing a test here)

the new expunger runs through all builds and does a rm-rf of everything
that is older than 24 months OR not one of the most recent 128 builds.
it then iterates again (somewhat inefficient, but meh) and mangles
all builds that are older than 6 months OR not one of the most 256 builds.

expunging has lower limit because 2 year old builds are suuuuper useless,
so we can be fairly aggressive (in point of fact, less may even still
be an option...).

mangling essentially defuncts the build WRT everything but the actual build
metadata  (iow logs and test data are all gone, along with artifacts)