restrict qt/framework dependee extension to unstable type jobs
(also refactor the awkward if to actually be readable)

the intent here is that Qt builds have dependees in unstable AND
stable AND release BUT only its unstable build.
with the introduction of staging repos (i.e. experimental) where the Qt
builds are staged before they migrate to unstable, the previous code was
insufficiently lax.

the revised code now also ensure that this dependee extension only gets
run for unstable jobs (i.e. unstable_qtbase will have release_foo as
dependee, but experimental_qtbase will not).

this also fixes a defect with framework detection caused by the
differentiation of project.component and .kdecomponent. unstable frameworks
would not have release as dependeeds but due to insufficient porting to
.kdecomponent this wasn't the case anymore. to prevent this from
happening the framework check has been deduplicated.