automatically set builds bin_only when their arch isn't the 'all' arch
this enables arm builds to build without Build-Depends-Indep and improves
Qt staging speed but also enables arm builds to perform faster when a
source supports bin-only builds.

this simply hijacks the dep_resolve bin_only value and forces it to true
so long as PANGEA_ARCH_BIN_ONLY isn't used to disable it.

bin_only_possible? is ignored since this should be good to go. if there
should be trouble we can always add in a whitelist/blacklist system later.

raising after build failures will now also report if a build was performed
bin-only (in the hopes of more easily being able to identify when
the bin-only may have something to do with the failure)