fix dpkg arching
undo part of 95d89d66a4f88c4aa7b4bd791a9420b721184961

that commit was flawed on multiple levels most of all the way it goes
about its business WRT run and the actual motivation for the change.

it introduced run_with_ec but that is utterly useless because build_binary
then was made to ues run_with_ec in a way that is simply a system() call.

to model this entire thing properly there is now a DPKG::Architecture class
which is specifically meant to carry state of a dpkg-architecture call.
this class then uses system to do the actual call

run_with_ec is now gone again and replaced by the class. this isolates
the system() call into a discrete unit we can test independently. this
allows the actual architecture moching to simply moch the architecture
object itself instead of the underlying calls