reduce code dupe and improve version handling of merginator
version handling is now centralized in a derivate of debian version which
provides additional splitting of the upstream version in its (likely)
real upstream and the debian-specific suffix (notably +dfsg).
this allows us to only pass epoch+realupstream into mangling functions
(where we don't want to mangle suffix or rev, becaues they are different
 across sources, so one is 5.13.2 and the other is 5.13.2+dfsg). it's
unclear if we actually need the epoch for the mangles, kinda depends on
whether a potential epoch would be introduced throughout Qt or only
modules that need them. time will tell

additionally the new version class allows us to only deal with this class
and mutate as necessary (for e.g. the new_version) without having to worry
about version building specifics anymore.

all this should yield greater stability when it comes to version handling

this leaves repo-handling with the greatest refactoring opportunity. i.e.
MerginatorRepo should operate on a single repo so it encapsulates the
state of that repo instead of having to have a huge ass run method so
as to have all the state method-local