add native test with useless override
when setting an upstream_scm override on a native packaging project that
override does nothing because native packages cannot have sources (this is
not a tooling opinion, but how dpkg works). as such the override makes
no sense and gets skipped over. previously this was done silently
which is nigh impossible to debug, so instead raise an error.

I have a bit of fear that this may break some existing stuff, after all
that nil check was there for a reason, since I have no clue what I guess
we'll have to deal with whatever fallout may come