skip over no-op overrides
because overrides cascade generic overrides could set stupid crap such
as uscan for everything. this makes no sense for native packages so they
need a way to opt out of this. easy way is to set the overridden thing null
which turns the operation no-op when the member is not default initialized
to anything