monkey patch tty command
still a bit awkward...

the default pretty printer of tty command is a bit ugh because it
forces each flushed string to be newlined and tab prefixed and colored.
this does cause the rather exiting problem that applications that flush
a lot (like apt; god knows why) can have single character lines. also
the way the tabbing works means that when flushing isn't done per-line
you'll have blocks of multiples lines where only the first line is tabbed
but the rest is not.
so what's the point even.

we now default to a new bespoke printer which does away with the \t and
the forced coloration and the forced newlining. instead the string is
output verbatim. we still keep the start and exit status information as
that is in part why I like to use tty command over system