throw out pxz
no longer a thing in focal. I am also not sure we actually use it
anywhere so let's drop it and deal with potential fallout (if any).

for source building I'm not sure it ever worked (properly). replace it
with --threads there. it's been a feature since a million years.

also move to tar+xz in one step since I'm already in the code