disable man-db trigger updates
+add new hook to generate a debconf preseed to do that

man-db triggers are fairly expensive and make apt installs take way longer
than they'd need to be. so, we disable them. the cache is only relevant
when you want to search for a manpage, which is very niche as of course
mans in general are, but even more so wanting to search for random words
within all mans.

the cache is still updated by man-db.timer which will be sufficient for
99.99999% of people (in fact, even that timer is already questionable, but
seeing as it doesn't do much other than use processing power I'm content
in leaving it enabled for now)