make nci use priv'd armhf for the time being
we are looking into other server options for the time being that server
is still stuck on xenial and errors out on utime syscalls. I suspect
it's simply the libseccomp not knowing what to do. in any event the only
way to get this sorted is by using priv'd containers (short of adjusting
the confinement which I would very much like to avoid unless absolutely
necessary as it is both more work and also possibly useless if we move
to a different server)
this isn't the hugest concern since we more or less only run trusted code
anyway