add job to prune old builds entirely
we do mangle builds but then we never throw them away. this is problematic
because jenkins runs on a hdd and iterating a backlog of 300 builds gets
slow quickly.
abandon builds entirely after 2 months. while more history is nice to have
there is like no use for it 99.999999% of the time