make version checks 3 million times faster
with the amount of packages we have apt show just doesn't scale anymore
it takes multiple seconds to get results and it's altogether horrible.
instead require a comprehensive list to be supplied once and then
resolve all apt-available packages in one call. this brings per
package cost down to sub 0.0X seconds at a negligible up-front
cost of maybe 10 seconds once.

to facilitate testing there's also a new class level cmd attribute
so we can mock only that. the dpkg version compares are all forks
so we'd mess with them if we'd stub TTY::Command as a whole.
a bit meh but nothing better comes to mind