implement lintian profile instead of bespoke parsing of lintian
previously we'd run lintian without modifications which results in it
creating a whole bunch of output in the log that we ultimately do
not care about. instead we now try to use a profile to get lintian
to not even generate output. this simplifies our linter as we now
only need to take care of some additional fiddling on top of lintian

whether or not this works as expected remains to be seen but codewise
it already seems much nicer than some of the hacks we pulled