make all lintian errors warnings
kdevelop is LITTERED with useless garbage. silence all of it should
anyone ever feel like something needs to be an error they can
make the mangler exclude it. but in 2020 the original system of errors
are actually meaningful means nothing for lintian and holding on to it
is a huge waste of time