rejigger adjust? to be more split between branch and url
as it turns out plasma stable builds are manually overridden to a
different branch which did trigger the original adjust? skip.
this no longer makes sense since we need to adjust the url for those
repos even though the branch is overridden.

the default_branch? is now only considered for branch overridability.
to mirror the same for url there's now default_url? which does much
the same but for url.

this should effectivley make the adjustment work for projects where the
override is either on branch (in which case we'll leave it untouched)
or url (in which case we'll leave that untouched but change branch).
the overall adjust condition is still that the url (which may have been
overridden, mind) looks like a kde gitty url