prevent anongit urls from being used ... at all!
there are ways that anongit may end up being used despite the adjustment
tech. notably overrides are not undone anymore which means if an override
uses anongit that'd be preserved and lead to bogus jobs

now since anongit is bound to die in a couple of days anyway there is ZERO
reason to have any urls be anongit.kde.org and so we now assert this fact
as part of the adjustment call. albeit regardless of the adjustment
actually having taken place to cover the aforementioned override edge case