force cmake internal build type for testing
only on release-lts and arm, latter by request from bshah.

the problem with the status quo is that dh set the CMAKE_BUILD_TYPE=None
and pkg-kde-tools (where used) sets it CMAKE_BUILD_TYPE=Debian in either
event they aren't actually valid types so cmake falls back to the global
flags which are dpkg-buildflags'.
that isn't a problem on its own but Qt's cmake configs only add -DNDEBUG
if and only if the type is releasy, which it never is because we set
invalid types.
so effectively we aren't fully releasy builds right now. since we don't
know what impact, if any, that has we are doing a test run with
selective builds switched to RelWithDebInfo (which effectively is the
same with the ability for Qt to set DNDEBUG correctly).

whether or not this is useful remains to be seen. a full on Release type
might make more sense overall in that it actually uses -O3 instead of
-O2 but it doesn't have debug symbols so from the POV that users may
want to report bugs for crashes that's not all that grand either :|